add vaccination courses

+++ DiplomCovid19/Models/EmployeeVaccineJunction.cs
@@ -7,8 +7,8 @@ namespace DiplomCovid19.Models
 {
     public class EmployeeVaccineJunction
     {
-        public int Id { get; set; }
-        public int? EmployeeId { get; set; }
+        public long Id { get; set; }
+        public long? EmployeeId { get; set; }
         public Employee Employee { get; set; }
         public int? VaccineId { get; set; }
         public Vaccine Vaccine { get; set; }
